requirements.txt
mysql-connector-python
sqlalchemy
pandas
numpy
